Don't install mpv if it already exists

+++ edgeware/EdgewareSetup.bat
@@ -14,8 +14,8 @@ goto checkPip
 echo Could not find Python.
 echo Now downloading installer from python.org, please wait...
 reg Query "HKLM\Hardware\Description\System\CentralProcessor\0" | find /i "x86" > NUL && set OS=32BIT || set OS=64BIT
-if %OS%==32BIT powershell -Command "(New-Object Net.WebClient).DownloadFile('https://www.python.org/ftp/python/3.12.6/python-3.12.6.exe', 'pyinstaller.exe')"
-if %OS%==64BIT powershell -Command "(New-Object Net.WebClient).DownloadFile('https://www.python.org/ftp/python/3.12.6/python-3.12.6-amd64.exe', 'pyinstaller.exe')"
+if %OS%==32BIT curl https://www.python.org/ftp/python/3.12.6/python-3.12.6.exe -o pyinstaller.exe
+if %OS%==64BIT curl https://www.python.org/ftp/python/3.12.6/python-3.12.6-amd64.exe -o pyinstaller.exe
 echo Done downloading executable.
 echo Please complete installation through the installer before continuing.
 start %CD%\pyinstaller.exe
@@ -42,13 +42,17 @@ py -m pip install -r requirements.txt
 if NOT %errorlevel%==0 goto quitRequirements
 goto mpv
 :mpv
-echo Installing libmpv...
-if not exist data mkdir data
-curl https://7-zip.org/a/7zr.exe -o data\7z.exe
-reg Query "HKLM\Hardware\Description\System\CentralProcessor\0" | find /i "x86" > NUL && set OS=32BIT || set OS=64BIT
-if %OS%==32BIT curl -L https://github.com/shinchiro/mpv-winbuild-cmake/releases/download/20250304/mpv-dev-i686-20250304-git-2542a78.7z -o data\mpv.7z
-if %OS%==64BIT curl -L https://github.com/shinchiro/mpv-winbuild-cmake/releases/download/20250304/mpv-dev-x86_64-20250304-git-2542a78.7z -o data\mpv.7z
-data\7z.exe e data\mpv.7z -odata libmpv-2.dll
+if not exist data\libmpv-2.dll (
+  echo Installing libmpv...
+  if not exist data mkdir data
+  if not exist data\7z.exe curl https://7-zip.org/a/7zr.exe -o data\7z.exe
+  if not exist data\mpv.7z (
+    reg Query "HKLM\Hardware\Description\System\CentralProcessor\0" | find /i "x86" > NUL && set OS=32BIT || set OS=64BIT
+    if %OS%==32BIT curl -L https://github.com/shinchiro/mpv-winbuild-cmake/releases/download/20250304/mpv-dev-i686-20250304-git-2542a78.7z -o data\mpv.7z
+    if %OS%==64BIT curl -L https://github.com/shinchiro/mpv-winbuild-cmake/releases/download/20250304/mpv-dev-x86_64-20250304-git-2542a78.7z -o data\mpv.7z
+  )
+  data\7z.exe e data\mpv.7z -odata libmpv-2.dll
+)
 goto shortcuts
 :shortcuts
 call :makePyw "CONFIG" "config"
